requirements.txt
aiomysql==0.2.0
anyio==4.4.0
certifi==2024.7.4
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
mysql-connector-python==9.0.0
PyMySQL==1.1.1
python-dotenv==1.0.1
python-telegram-bot==21.3
sniffio==1.3.1
typing_extensions==4.12.2
